# API-Spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Employee Task Management REST API",
    "version": "1.0.0",
    "description": "REST API documentation for Employee Task Management application with JWT security."
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/employees/{id}": {
      "get": {
        "summary": "Get employee by ID",
        "tags": [
          "employee-controller"
        ],
        "operationId": "getEmployeeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                },
                "examples": {
                  "viewEmployeeByIdSuccess": {
                    "summary": "View Employee By Id - Success",
                    "value": {
                      "id": 14,
                      "name": "Hania Ahmad",
                      "email": "whohaniadsd12@gmail.com",
                      "role": "Hania Ahmad"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "viewEmployeeByIdInvalidToken": {
                    "summary": "View Employee By Id - Invalid Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T15:56:44.675308"
                    }
                  },
                  "viewEmployeeByIdNoToken": {
                    "summary": "View Employee By Id - No Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T15:54:19.308195"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "viewEmployeeByIdEmployeeAccessDenied": {
                    "summary": "View Employee By Id - Employee Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-06T16:07:02.837292"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "getEmployeeByIdNotFound": {
                    "summary": "Get Employee By ID - User Not Found",
                    "value": {
                      "message": "User Not Found",
                      "status": 404,
                      "timestamp": "2026-08-06T15:47:30.522364"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Resource conflict or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "viewEmployeeByIdUserNotFound": {
                    "summary": "View Employee By Id - User Not Found",
                    "value": {
                      "message": "User Not Found",
                      "status": 409,
                      "timestamp": "2026-08-06T15:47:30.522364"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update employee details",
        "tags": [
          "employee-controller"
        ],
        "operationId": "updateEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                },
                "examples": {
                  "updateEmployeeSuccess": {
                    "summary": "Update Employee \u2013 Success",
                    "value": {
                      "id": 15,
                      "name": "John Wick",
                      "email": "johh@gmail.com",
                      "role": "EMPLOYEE"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidEmailFormat": {
                    "summary": "Invalid Email Format",
                    "value": {
                      "message": "Invalid email format",
                      "status": 400,
                      "timestamp": "2026-08-06T18:36:31.780541"
                    }
                  },
                  "emptyEmail": {
                    "summary": "Empty Email",
                    "value": {
                      "message": "Email cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T18:35:26.340668"
                    }
                  },
                  "emptyName": {
                    "summary": "Empty Name",
                    "value": {
                      "message": "Name cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T18:34:38.619514"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidToken": {
                    "summary": "Invalid Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T18:39:08.066583"
                    }
                  },
                  "noToken": {
                    "summary": "No Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T18:37:56.007883"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "employeeAccessDenied": {
                    "summary": "Employee Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-06T18:40:14.804908"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "updateEmployeeNotFound": {
                    "summary": "Update Employee - User Not Found",
                    "value": {
                      "message": "User Not Found",
                      "status": 404,
                      "timestamp": "2026-08-06T18:32:20.22668"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Resource conflict or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "duplicateEmail": {
                    "summary": "Duplicate Email",
                    "value": {
                      "message": "Email already exists",
                      "status": 409,
                      "timestamp": "2026-08-06T18:33:26.862069"
                    }
                  },
                  "userNotFound": {
                    "summary": "User Not Found",
                    "value": {
                      "message": "User Not Found",
                      "status": 409,
                      "timestamp": "2026-08-06T18:32:20.22668"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete employee",
        "tags": [
          "employee-controller"
        ],
        "operationId": "deleteEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                },
                "examples": {
                  "deleteEmployeeSuccessfully": {
                    "summary": "Delete Employee Successfully",
                    "value": {
                      "message": "Employee deleted successfully",
                      "timestamp": "2026-08-06T19:04:09"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidToken": {
                    "summary": "Invalid Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T19:11:12.486206"
                    }
                  },
                  "noToken": {
                    "summary": "No Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T19:10:08.76878"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role or admin deletion forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "deleteEmployeeSuccessfully": {
                    "summary": "Delete Employee Successfully",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-06T19:15:20.767833"
                    }
                  },
                  "deleteAdminNotAllowed": {
                    "summary": "Delete Admin Not Allowed",
                    "value": {
                      "message": "Admin cannot be deleted",
                      "status": 403,
                      "timestamp": "2026-08-06T19:06:53.249775"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "deleteEmployeeNotFound": {
                    "summary": "Delete Employee - User Not Found",
                    "value": {
                      "message": "User Not Found",
                      "status": 404,
                      "timestamp": "2026-08-06T19:05:43.966663"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Resource conflict or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "userNotFound": {
                    "summary": "User Not Found",
                    "value": {
                      "message": "User Not Found",
                      "status": 409,
                      "timestamp": "2026-08-06T19:05:43.966663"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/employees": {
      "get": {
        "summary": "Get all employees",
        "tags": [
          "employee-controller"
        ],
        "operationId": "getAllEmployees",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeResponse"
                  }
                },
                "examples": {
                  "viewAllEmployeesSuccessAdminToken": {
                    "summary": "View All Employees - Success (Admin token)",
                    "value": [
                      {
                        "id": 3,
                        "name": "Amash Ahmad",
                        "email": "amash@gmail.com",
                        "role": "EMPLOYEE"
                      },
                      {
                        "id": 4,
                        "name": "Hania Ahmad",
                        "email": "hania@gmail.com",
                        "role": "EMPLOYEE"
                      },
                      {
                        "id": 10,
                        "name": "John Doe",
                        "email": "john@gmail.com",
                        "role": "EMPLOYEE"
                      },
                      {
                        "id": 11,
                        "name": "Joh",
                        "email": "joh@gmail.com",
                        "role": "EMPLOYEE"
                      }
                    ]
                  },
                  "viewAllEmployeesSuccessAdminTokenNoEmployee": {
                    "summary": "View All Employees - Success (Admin token) No Employee",
                    "value": [
                      {
                        "id": 3,
                        "name": "Amash Ahmad",
                        "email": "amash@gmail.com",
                        "role": "EMPLOYEE"
                      },
                      {
                        "id": 4,
                        "name": "Hania Ahmad",
                        "email": "hania@gmail.com",
                        "role": "EMPLOYEE"
                      },
                      {
                        "id": 10,
                        "name": "John Doe",
                        "email": "john@gmail.com",
                        "role": "EMPLOYEE"
                      },
                      {
                        "id": 11,
                        "name": "Joh",
                        "email": "joh@gmail.com",
                        "role": "EMPLOYEE"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "viewAllEmployeesInvalidToken": {
                    "summary": "View All Employees - Invalid token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T12:52:56.578528"
                    }
                  },
                  "viewAllEmployeesNoToken": {
                    "summary": "View All Employees - No token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T12:51:51.735019"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "viewAllEmployeesEmployeeTokenAccessDenied": {
                    "summary": "View All Employees - Employee token Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-06T14:13:06.159028"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new employee",
        "tags": [
          "employee-controller"
        ],
        "operationId": "createEmployee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                },
                "examples": {
                  "createEmployeeByAdminSuccess": {
                    "summary": "Create Employee - By Admin Success",
                    "value": {
                      "id": 11,
                      "name": "Joh",
                      "email": "joh@gmail.com",
                      "role": "EMPLOYEE"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createEmployeeShortPassword": {
                    "summary": "Create Employee - Short Password",
                    "value": {
                      "message": "Password must be at least 8 characters",
                      "status": 400,
                      "timestamp": "2026-08-06T11:52:24.055892"
                    }
                  },
                  "createEmployeeEmptyPassword": {
                    "summary": "Create Employee - Empty Password",
                    "value": {
                      "message": "Password cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T11:50:52.588675"
                    }
                  },
                  "createEmployeeInvalidEmailFormat": {
                    "summary": "Create Employee - Invalid Email Format",
                    "value": {
                      "message": "Invalid email format",
                      "status": 400,
                      "timestamp": "2026-08-06T11:48:26.048088"
                    }
                  },
                  "createEmployeeEmptyEmail": {
                    "summary": "Create Employee - Empty Email",
                    "value": {
                      "message": "Email cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T11:46:11.056449"
                    }
                  },
                  "createEmployeeEmptyName": {
                    "summary": "Create Employee - Empty Name",
                    "value": {
                      "message": "Name cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T11:43:18.978131"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createEmployeeInvalidToken": {
                    "summary": "Create Employee - Invalid Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T11:56:31.050204"
                    }
                  },
                  "createEmployeeNoToken": {
                    "summary": "Create Employee - No Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T11:55:13.708823"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createEmployeeByEmployeeAccessDenied": {
                    "summary": "Create Employee - By Employee Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-06T10:32:23.256834"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Employee with given email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createEmployeeDuplicateEmail": {
                    "summary": "Create Employee - Duplicate Email",
                    "value": {
                      "message": "Employee Already Exist",
                      "status": 409,
                      "timestamp": "2026-08-06T11:40:00.345107"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdateEmployeeRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "name"
        ]
      },
      "EmployeeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "EmployeeRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
...
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}